requirements.txt
# Django Core
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-channels>=4.0
channels-redis>=4.2

# MongoDB
djongo>=1.3.6
pymongo>=4.6,<5.0

# Task Queue
celery>=5.3
redis>=5.0

# Broker APIs
alpaca-trade-api>=3.2
ccxt>=4.2

# Security
python-dotenv>=1.0
cryptography>=42.0

# HTTP / Utilities
requests>=2.31
gunicorn>=21.2

# Testing
factory-boy>=3.3
